Use wrapping subtraction when dealing with align masks
A common pattern when applying an align mask is to subtract 1 from
the value to be aligned, OR it with the align mask, and add back 1
again.

This works fine when the value is zero, but does rely on the subtraction
and addition to wrap, which Rust does not permit by default.

So use wrapping_sub()/add() in these cases.

Fixes: #2