Take non-zero image base into account when applying relocations
Gerd reports that UKIs, which have a non-zero ImageBase, are not loaded
correctly. This is due to the fact that all relocated quantities already
take the image base into account, and so applying the relocations
involves applying the delta between the build time and actual image
base.

This also means that there is no need to apply the relocations at all
when the actual image base equals the build time image base, but this
can be fixes another time.

Fix proposed by Gerd Hoffman (@kraxel)

Fixes: #5